Move embedding exceptions into their own namespace
